refactor(trade_executor): Add error handling to trade executor impl
